show need not found; refactoring
